Rename components for staff
