+++ src/components/Staff/NavBar.vue
@@ -6,11 +6,11 @@
         <span class="ml-2 orangeText">FDS</span>
       </div>
     </router-link>
-    <router-link to="/staff/order-summary">
-            <v-btn class="ml-10 orange--text text--darken-1" outlined>Order Statistics</v-btn>
+    <router-link to="/staff/add-promo">
+        <v-btn class="ml-10 orange--text text--darken-1" outlined>Add Promotion</v-btn>
     </router-link>
-     <router-link to="/staff/promo-summary">
-            <v-btn class="ml-10 orange--text text--darken-1" outlined>Promotion Statistics</v-btn>
+    <router-link to="/staff/view-summary">
+        <v-btn class="ml-10 orange--text text--darken-1" outlined>View Summary</v-btn>
     </router-link>
     <v-spacer></v-spacer>
     <v-btn outlined color="orange" @click="logout">Logout</v-btn>
