+++ src/router/index.js
@@ -5,8 +5,8 @@ import Customer from '@/components/Customer/Customer'
 import CustomerBrowse from '@/components/Customer/Browse'
 import Rider from '@/components/Rider/Rider'
 import Staff from '@/components/Staff/Staff'
-import OrderSummary from '@/components/Staff/OrderSummary'
-import PromotionSummary from '@/components/Staff/PromotionSummary'
+import ViewSummary from '@/components/Staff/ViewSummary'
+import AddPromotion from '@/components/Staff/AddPromotion'
 import Manager from '@/components/Manager/Manager'
 import ErrorLanding from '@/components/ErrorLanding'
 import store from '../store'
@@ -59,8 +59,8 @@ const router = new VueRouter({
             }
         },
         {
-            path: '/staff/order-summary',
-            component: OrderSummary,
+            path: '/staff/view-summary',
+            component: ViewSummary,
             beforeEnter: (to, from, next) => {
                 if (store.getters.role != 'Staff') {
                     router.push({ name: 'ErrorLanding' })
@@ -69,8 +69,8 @@ const router = new VueRouter({
             }
         },
         {
-            path: '/staff/promo-summary',
-            component: PromotionSummary,
+            path: '/staff/add-promo',
+            component: AddPromotion,
             beforeEnter: (to, from, next) => {
                 if (store.getters.role != 'Staff') {
                     router.push({ name: 'ErrorLanding' })
